Don't bundle-build lan server either
