+++ packages/lan/package.json
@@ -13,12 +13,19 @@
     "node": "^16.16.0"
   },
   "scripts": {
-    "start": "node ./dist/app.bundle.js",
-    "start-dev": "rimraf ./dist/ && webpack --config webpack.config.dev.mjs",
+    "build": "swc --delete-dir-on-start --out-dir dist --copy-files --source-maps true app",
+    "start": "node dist",
+
     "test": "NODE_ENV=test jest",
+    "lint": "eslint app",
+
+    "setup-dev": "node dist setup",
+    "shell": "yarn build && node dist shell",
+
+    "start-watch": "chokidar \"dist/**/*.js\" -c \"yarn run start\"",
     "test-watch": "yarn test --watch",
-    "build": "rimraf ./dist/ && webpack --config webpack.config.prod.mjs",
-    "lint": "eslint app"
+    "build-watch": "yarn run build --watch",
+    "start-dev": "concurrently npm:build-watch npm:start-watch"
   },
   "devDependencies": {
     "@beyondessential/eslint-config-js": "^1.1.0",
@@ -26,7 +33,6 @@
     "@swc/jest": "0.2.26",
     "chai": "^4.1.2",
     "eslint": "^5.15.2",
-    "eslint-import-resolver-webpack": "^0.11.0",
     "eslint-plugin-import": "^2.22.1",
     "fast-check": "^3.3.0",
     "jest": "^29.3.1",
