--- packages/lan/start.bat
@@ -1,3 +1,0 @@
-:: Windows batch file to help users retrieve debug output
-server.exe
-pause
