--- packages/lan/webpack.config.prod.mjs
@@ -1,6 +1,0 @@
-import { NODE_WEBPACK_CONFIG_SWC } from '@tamanu/build-tooling';
-
-export default {
-  ...NODE_WEBPACK_CONFIG_SWC,
-  mode: 'production',
-};
